% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chiliconcarne.R
\docType{package}
\name{Chiliconcarne}
\alias{Chiliconcarne}
\alias{Chiliconcarne-package}
\title{A package created for a series of labs for a course in Advanced R-programming}
\description{
A package created for a series of labs for a course in Advanced R-programming
}

