% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{Computes the greatest common devisor of two numeric integers via
the Euclidian algorithm. The implementation is division based and uses the
modulus function.}
\usage{
euclidian(s, b)
}
\description{
Computes the greatest common devisor of two numeric integers via
the Euclidian algorithm. The implementation is division based and uses the
modulus function.
}

